pyproject.toml
[tool.pyright]
include = ["rippling"]
executionEnvironments = [
  { root = "rippling" }
]
pythonVersion = "3.10"
venvPath = "./"
venv = ".ddd"

[tool.poetry]
name = "ddd-demo"
version = "0.1.0"
description = ""
authors = ["Flo Robert <frankie.robert+github@gmail.com>"]
readme = "README.md"
packages = [{include = "ddd_demo"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.7"
django-injector = "^0.3.1"
django-mongoengine = "^0.5.6"
pymongo = "^4.6.0"
djangorestframework = "^3.14.0"
djangorestframework-stubs = "^3.14.4"
drf-typed = "^0.3.0"
pydantic = "^2.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
